Material Cockpit v2 — move the jump processes into de.metas.ui.web.base
These are WebUI processes, so they belong beside the rest of the cockpit-v2 WebUI
code rather than in de.metas.swat.base. They move to
de.metas.ui.web.material.cockpit.v2.jump, alongside the sibling
…cockpit.v2.reservation package, which already holds an abstract base plus its
concrete processes. The three pre-existing jump classes move with the two this
work added, so the family stays together. No POM change: de.metas.ui.web.base
already reaches de.metas.inoutcandidate and de.metas.material.cockpit and depends
on de.metas.manufacturing directly.

The move removes a duplication this work had documented as unavoidable. The base
declared its own MSG_NO_RELATED_DOCS_FOUND only because de.metas.swat.base cannot
see SqlViewFactory — the dependency runs the other way. From
de.metas.ui.web.base it simply uses SqlViewFactory.MSG_NO_RELATED_DOCS_FOUND, the
same constant RelationTypeInOverlayProcess uses; same AD_Message, same behaviour.

AD_Process.Classname stores the fully-qualified name, so the migration updates it
for the three processes; without it the WebUI could not instantiate them and every
jump action would disappear. It keys on the old classname as well as the id, so a
re-run cannot clobber a corrected row, and it leaves 585515 alone — that one runs
the shared RelationTypeInOverlayProcess, which is not moving.
