+++ backend/de.metas.ui.web.base/src/main/java/de/metas/ui/web/material/cockpit/v2/jump/QtyDemand_QtySupply_V_to_PP_Order_Candidate.java
@@ -20,7 +20,7 @@
  * #L%
  */
 
-package de.metas.material.process;
+package de.metas.ui.web.material.cockpit.v2.jump;
 
 import de.metas.material.cockpit.QtyDemandQtySupply;
 import lombok.NonNull;
